💥  rehaul `getStaticPropsCatchAll` ✨️ Podcast|Preview|Clear
✨️ Preview & Clear Functionality finalized
✨️ `getPodcastFeed` introduced to share that across multi-site

💥️ Then the rehaul of how we use `getStaticPropsCatchAll`
- No longer pass `clear|pathVariables`
- New export is `{ data, pathVariables }`

🏷️ Types for `image/...` key via `IGetPlaiceholderReturn`

BREAKING CHANGE: 💥️ Alter `slug` to `id` for `image/...` keys, complete rehaul of getStaticPropsCatchAll
