+++ packages/shared/src/lib/types/image.types.ts
@@ -1,8 +1,9 @@
 import type { IGetPlaiceholderReturn } from 'plaiceholder'
+import type { IGetBlurhashReturn } from 'plaiceholder/dist/blurhash'
 
 interface IGetPlaiceholderReturnCustom
   extends Omit<IGetPlaiceholderReturn, 'blurhash' | 'css' | 'svg'> {
-  blurhash?: string
+  blurhash?: IGetBlurhashReturn
   id: string
   url: string
 }
