+++ sites/jeromefitzgerald.com/src/pages/api/v1/cms/preview-clear.ts
@@ -1,0 +1,8 @@
+// eslint-disable-next-line @typescript-eslint/require-await
+const PreviewClear = async (_, res) => {
+  res.clearPreviewData()
+  res.writeHead(307, { Location: '/' })
+  res.end()
+}
+
+export default PreviewClear
