+++ sites/jeromefitzgerald.com/src/routes/Episodes/Card.tsx
@@ -18,6 +18,7 @@ import {
 } from '@jeromefitz/design-system/components/Card/Show'
 import type { Episode as EpisodeProperties } from '@jeromefitz/notion/schema'
 import { IMAGE__PLACEHOLDER } from '@jeromefitz/shared/src/lib/constants'
+import type { IGetPlaiceholderReturnCustom } from '@jeromefitz/shared/src/lib/types'
 import { lpad } from '@jeromefitz/utils'
 import _isEmpty from 'lodash/isEmpty'
 import { getNotion } from 'next-notion/src/helper'
@@ -91,13 +92,13 @@ const EpisodesCard = ({
    */
   const imageSlug = !!seoImage && Object.keys(seoImage)[0]
   const url = !!imageSlug && seoImage[imageSlug]?.url
-  const fallbackData =
+  const fallbackData: IGetPlaiceholderReturnCustom =
     !!url && !!images && !!imageSlug
       ? images[`image/${imageSlug}`]
       : {
           base64: IMAGE__PLACEHOLDER.base64,
+          id: IMAGE__PLACEHOLDER.id,
           img: { ...IMAGE__PLACEHOLDER.img, src: url },
-          slug: IMAGE__PLACEHOLDER.slug,
         }
   // @note(image) limit api calls
   // @refactor(lodash) _isEmpty
@@ -105,8 +106,8 @@ const EpisodesCard = ({
   // console.dir(fallbackData)
   const urlApi = !!url && _isEmpty(fallbackData) ? `/api/v1/img?url=${url}` : null
   const {
-    data: { base64, img, slug: slugImage },
-  } = useSWR<any>(urlApi, fetcher, {
+    data: { base64, id: slugImage, img },
+  } = useSWR<IGetPlaiceholderReturnCustom>(urlApi, fetcher, {
     fallbackData,
   })
