+++ sites/jeromefitzgerald.com/src/routes/Shows/Card.tsx
@@ -17,6 +17,7 @@ import {
 } from '@jeromefitz/design-system/components/Card/Show'
 import type { Show } from '@jeromefitz/notion/schema'
 import { IMAGE__PLACEHOLDER } from '@jeromefitz/shared/src/lib/constants'
+import type { IGetPlaiceholderReturnCustom } from '@jeromefitz/shared/src/lib/types'
 import _isEmpty from 'lodash/isEmpty'
 import _map from 'lodash/map'
 import { fetcher } from 'next-notion/src/lib/fetcher'
@@ -71,13 +72,13 @@ const ShowsCard = ({
    */
   const imageSlug = !!seoImage && Object.keys(seoImage)[0]
   const url = !!imageSlug && seoImage[imageSlug]?.url
-  const fallbackData =
+  const fallbackData: IGetPlaiceholderReturnCustom =
     !!url && !!images && !!imageSlug
       ? images[`image/${imageSlug}`]
       : {
           base64: IMAGE__PLACEHOLDER.base64,
+          id: IMAGE__PLACEHOLDER.id,
           img: { ...IMAGE__PLACEHOLDER.img, src: url },
-          slug: IMAGE__PLACEHOLDER.slug,
         }
   // @note(image) limit api calls
   // @refactor(lodash) _isEmpty
@@ -88,8 +89,8 @@ const ShowsCard = ({
   // console.dir(imageSlug)
   const urlApi = !!url && _isEmpty(fallbackData) ? `/api/v1/img?url=${url}` : null
   const {
-    data: { base64, img, slug: slugImage },
-  } = useSWR<any>(urlApi, fetcher, {
+    data: { base64, id: slugImage, img },
+  } = useSWR<IGetPlaiceholderReturnCustom>(urlApi, fetcher, {
     fallbackData,
   })
